+++ apex/amp/_process_optimizer.py
@@ -3,7 +3,7 @@ from ..fp16_utils import master_params_to_model_params
 from ..multi_tensor_apply import multi_tensor_applier
 from ._amp_state import maybe_print
 import torch
-from ..optimizers import FusedAdam, FusedSGD
+from ..optimizers import FusedSGD
 
 
 class AmpOptimizerState(object):
@@ -242,66 +242,7 @@ def post_backward_no_master_weights(self, scaler):
 
 
 #####################################################################################
-# FusedAdam versions
-#####################################################################################
-
-def prepare_backward_with_master_weights_FusedAdam(self):
-    stash = self._amp_stash
-
-    self._amp_lazy_init()
-
-
-def post_backward_with_master_weights_FusedAdam(self, scaler):
-    stash = self._amp_stash
-
-    self._amp_lazy_init()
-
-    stash.scale = scaler.loss_scale()
-    stash.grads = [[param.grad.data for param in group] for group in stash.fp16_groups]
-    stash.output_params = [[param for param in group] for group in stash.fp16_groups]
-
-    norm_groups = []
-    skip = False
-    for grad_group in stash.grads:
-        norm, _ = multi_tensor_applier(
-            stash.multi_tensor_l2norm,
-            stash.dummy_overflow_buf,
-            [grad_group],
-            False)
-        # Still syncing here for now.
-        norm = float(norm)
-        norm_groups.append(norm)
-        if norm == float('inf') or norm == -float('inf') or norm != norm:
-            skip = True
-    if skip:
-        scaler._overflow_buf.fill_(1.)
-        scaler._has_overflow = True
-
-    stash.grad_norms = norm_groups
-
-
-def prepare_backward_no_master_weights_FusedAdam(self):
-    stash = self._amp_stash
-
-    self._amp_lazy_init()
-
-
-def post_backward_no_master_weights_FusedAdam(self, scaler):
-    stash = self._amp_stash
-
-    self._amp_lazy_init()
-
-    stash.scale = scaler.loss_scale()
-    stash.grads = None
-    stash.output_params = None
-    stash.grad_norms = None
-
-
-#####################################################################################
 # FusedSGD versions
-# Eat this ugly code duplication for now.  First make it work, then make it clean.
-# It's difficult to anticipate what can be unified between the FusedAdam and FusedSGD
-# implementations until I have them both working.
 #####################################################################################
 
 # FusedSGD never explicitly materializes the fp32 gradients for "fp32 from fp16" master params
@@ -406,7 +347,7 @@ def _process_optimizer(optimizer, properties):
             if closure is not None:
                 raise RuntimeError("Currently, Amp does not support closure use with optimizers.")
             retval = old_step()
-            if not (isinstance(self, FusedAdam) or isinstance(self, FusedSGD)):
+            if not isinstance(self, FusedSGD):
                 self._master_params_to_model_params()
             # Clear the master grads that wouldn't be zeroed by model.zero_grad()
             for param in self._amp_stash.all_fp32_from_fp16_params:
@@ -432,12 +373,7 @@ def _process_optimizer(optimizer, properties):
                 param.grad = None
         optimizer.zero_grad = types.MethodType(new_zero_grad, optimizer)
 
-        if isinstance(optimizer, FusedAdam):
-            optimizer._prepare_amp_backward = types.MethodType(
-                prepare_backward_with_master_weights_FusedAdam, optimizer)
-            optimizer._post_amp_backward = types.MethodType(
-                post_backward_with_master_weights_FusedAdam, optimizer)
-        elif isinstance(optimizer, FusedSGD):
+        if isinstance(optimizer, FusedSGD):
             optimizer._prepare_amp_backward = types.MethodType(
                 prepare_backward_with_master_weights_FusedSGD, optimizer)
             optimizer._post_amp_backward = types.MethodType(
@@ -451,12 +387,7 @@ def _process_optimizer(optimizer, properties):
         optimizer._lazy_init_maybe_master_weights = types.MethodType(
             lazy_init_no_master_weights, optimizer)
 
-        if isinstance(optimizer, FusedAdam):
-            optimizer._prepare_amp_backward = types.MethodType(
-                prepare_backward_no_master_weights_FusedAdam, optimizer)
-            optimizer._post_amp_backward = types.MethodType(
-                post_backward_no_master_weights_FusedAdam, optimizer)
-        elif isinstance(optimizer, FusedSGD):
+        if isinstance(optimizer, FusedSGD):
             optimizer._prepare_amp_backward = types.MethodType(
                 prepare_backward_no_master_weights_FusedSGD, optimizer)
             optimizer._post_amp_backward = types.MethodType(
