+++ apex/optimizers/__init__.py
@@ -1,6 +1,4 @@
 from .fused_sgd import FusedSGD
 from .fused_adam import FusedAdam
+from .fused_novograd import FusedNovoGrad
 from .fp16_optimizer import FP16_Optimizer
-from .sgd import SGD
-from .adam import Adam
-from .novograd import NovoGrad
