--- apex/optimizers/sgd.py
@@ -1,180 +1,0 @@
-import torch
-from apex.multi_tensor_apply import multi_tensor_applier
-
-class SGD(torch.optim.Optimizer):
-    r"""Implements stochastic gradient descent (optionally with momentum).
-    Nesterov momentum is based on the formula from
-    `On the importance of initialization and momentum in deep learning`__.
-    Args:
-        params (iterable): iterable of parameters to optimize or dicts defining
-            parameter groups
-        lr (float): learning rate
-        momentum (float, optional): momentum factor (default: 0)
-        weight_decay (float, optional): weight decay (L2 penalty) (default: 0)
-        dampening (float, optional): dampening for momentum (default: 0)
-        nesterov (bool, optional): enables Nesterov momentum (default: False)
-        set_grad_none (bool, optional): whether set grad to None when zero_grad()
-            method is called. (default: True)
-    Example:
-        >>> optimizer = torch.optim.SGD(model.parameters(), lr=0.1, momentum=0.9)
-        >>> optimizer.zero_grad()
-        >>> loss_fn(model(input), target).backward()
-        >>> optimizer.step()
-    __ http://www.cs.toronto.edu/%7Ehinton/absps/momentum.pdf
-    .. note::
-        The implementation of SGD with Momentum/Nesterov subtly differs from
-        Sutskever et. al. and implementations in some other frameworks.
-        Considering the specific case of Momentum, the update can be written as
-        .. math::
-                  v = \rho * v + g \\
-                  p = p - lr * v
-        where p, g, v and :math:`\rho` denote the parameters, gradient,
-        velocity, and momentum respectively.
-        This is in contrast to Sutskever et. al. and
-        other frameworks which employ an update of the form
-        .. math::
-             v = \rho * v + lr * g \\
-             p = p - v
-        The Nesterov version is analogously modified.
-    """
-
-    def __init__(self, params, lr=0.1, momentum=0., dampening=0.,
-                 weight_decay=0., nesterov=False, wd_after_momentum=False,
-                 set_grad_none=True):
-        if lr < 0.0:
-            raise ValueError("Invalid learning rate: {}".format(lr))
-        if momentum < 0.0:
-            raise ValueError("Invalid momentum value: {}".format(momentum))
-        if weight_decay < 0.0:
-            raise ValueError("Invalid weight_decay value: {}".format(weight_decay))
-
-        defaults = dict(lr=lr, momentum=momentum, dampening=dampening,
-                        weight_decay=weight_decay, nesterov=nesterov)
-
-        self.wd_after_momentum = wd_after_momentum
-        self.set_grad_none = set_grad_none
-        if multi_tensor_applier.available:
-            import amp_C
-            # Skip buffer
-            self._dummy_overflow_buf = torch.cuda.IntTensor([0])
-            self.multi_tensor_axpby = amp_C.multi_tensor_axpby
-            self.multi_tensor_sgd = amp_C.multi_tensor_sgd
-        else:
-            raise RuntimeError('apex.optimizers.SGD requires cuda extensions')
-
-        if nesterov and (momentum <= 0 or dampening != 0):
-            raise ValueError("Nesterov momentum requires a momentum and zero dampening")
-        super(SGD, self).__init__(params, defaults)
-
-    def __setstate__(self, state):
-        super(SGD, self).__setstate__(state)
-        for group in self.param_groups:
-            group.setdefault('nesterov', False)
-
-    def zero_grad(self):
-        if self.set_grad_none:
-            for group in self.param_groups:
-                for p in group['params']:
-                    p.grad = None
-        else:
-            super(SGD, self).zero_grad()
-
-    def get_momentums(self, params):
-        momentums = []
-        first_run = True
-        for p in params:
-            param_state = self.state[p]
-            # torch.optim.SGD initializes momentum in the main loop, we have
-            # to do it here, and track whether or not we've done so, so that
-            # momentum application can be skipped in the main kernel.
-            if 'momentum_buffer' not in param_state:
-                first_run = True
-                buf = param_state['momentum_buffer'] = torch.zeros_like(p.data)
-                momentums.append(buf)
-            else:
-                first_run = False
-                momentums.append(param_state['momentum_buffer'])
-        return momentums, first_run
-
-    def step(self, closure=None):
-        """Performs a single optimization step.
-        Arguments:
-            closure (callable, optional): A closure that reevaluates the model
-                and returns the loss.
-        """
-        loss = None
-        if closure is not None:
-            loss = closure()
-
-        for group in self.param_groups:
-            weight_decay = group['weight_decay']
-            momentum = group['momentum']
-            dampening = group['dampening']
-            nesterov = group['nesterov']
-
-            for group_dtype in [torch.float16, torch.float32]:
-                grad_list = [p.grad for p in group['params'] if (p.dtype == group_dtype and p.grad is not None)]
-                if len(grad_list) == 0:
-                    continue
-                param_list = [p for p in group['params'] if (p.dtype == group_dtype and p.grad is not None)]
-                if momentum != 0:
-                    momentum_list, first_run = self.get_momentums(param_list)
-                    multi_tensor_applier(
-                        self.multi_tensor_sgd,
-                        self._dummy_overflow_buf,
-                        [grad_list, param_list, momentum_list],
-                        weight_decay,
-                        momentum,
-                        dampening,
-                        group['lr'],
-                        nesterov,
-                        first_run,
-                        self.wd_after_momentum,
-                        1.0)
-                else:
-                    # show how to implement SGD using axpby, without writing new multi_tensor kernel
-                    # only enabled now in no momentum case, since it saves creating momentum for us
-                    # keep momentum != 0 code below for completeness
-                    if weight_decay != 0 and not self.wd_after_momentum:
-                        multi_tensor_applier(
-                            self.multi_tensor_axpby,
-                            self._dummy_overflow_buf,
-                            [grad_list, param_list, grad_list],
-                            1.,
-                            weight_decay,
-                            2)
-                    if momentum != 0: # always False
-                        if not first_run:
-                            multi_tensor_applier(
-                                self.multi_tensor_axpby,
-                                self._dummy_overflow_buf,
-                                [momentum_list, grad_list, momentum_list],
-                                momentum,
-                                1.-dampening,
-                                2)
-                        if nesterov:
-                            multi_tensor_applier(
-                                self.multi_tensor_axpby,
-                                self._dummy_overflow_buf,
-                                [grad_list, momentum_list, grad_list],
-                                1.,
-                                momentum,
-                                2)
-                        else:
-                            grad_list = momentum_list
-                    if weight_decay != 0 and self.wd_after_momentum:
-                        multi_tensor_applier(
-                            self.multi_tensor_axpby,
-                            self._dummy_overflow_buf,
-                            [grad_list, param_list, grad_list],
-                            1.,
-                            weight_decay,
-                            2)
-                    multi_tensor_applier(
-                        self.multi_tensor_axpby,
-                        self._dummy_overflow_buf,
-                        [param_list, grad_list, param_list],
-                        1.,
-                        -group['lr'],
-                        2)
-        return loss
