+++ tests/L0/run_test.py
@@ -1,7 +1,7 @@
 import unittest
 import sys
 
-test_dirs = ["run_amp", "run_fp16util", "run_mixed_adam", "run_fused_layer_norm"]
+test_dirs = ["run_amp", "run_fp16util", "run_fused_layer_norm", "run_optimizers"]
 
 runner = unittest.TextTestRunner(verbosity=2)
